Filters out NaN values from the db
That's the big change.  Trying to get running mean with nans in there
doesn't work.  It's best to just pull the meaningless rows from the
system; I learned that in the CLI version but didn't automatically do it
that way again.  Because I'm stupid.

Stuff for pulling the 1st transit from the NASA data is in here too.